requirements.txt
requests>=2.31.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
jinja2>=3.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
schedule>=1.2.0
weasyprint>=60.0
